requirements.txt
# Data manipulation
pandas==2.1.4
numpy==1.26.2
pyarrow==14.0.1

# Data acquisition
kaggle==1.5.16
requests==2.31.0

# Cloud storage
boto3==1.34.10
google-cloud-storage==2.14.0

# Data validation
great-expectations==0.18.8

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1
tqdm==4.66.1

# Database
duckdb==0.9.2
psycopg2-binary==2.9.9
sqlalchemy==2.0.23

# Jupyter for exploration
jupyter==1.0.0
ipykernel==6.27.1

# Machine learning
scikit-learn==1.3.2
xgboost==2.0.3
mlflow==2.9.2
matplotlib==3.8.2

# API
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3

# Routing
googlemaps==4.10.0
polyline==2.0.2